opticast-mining/requirements.txt
streamlit
google-generativeai
pandas
xlsxwriter
openpyxl